Split memoryService.ts core-memory loaders out (664 → 89 lines)
- memoryService.ts                MemoryService class — CRUD methods
                                  + thin wrappers for the core-memory
                                  initialization paths
- memory/types.ts                 CoreMemoryConfig interface
- memory/fallbackConfig.ts        FALLBACK_CORE_MEMORY_CONFIG — the
                                  in-memory default used when
                                  core.memory.json is missing/invalid
- memory/loadFromFile.ts          loadCoreMemoryFromFile + serialize +
                                  upsertCoreSection + validation. Takes
                                  an onFallback callback to keep the
                                  fallback dependency loose.
- memory/initializeDefault.ts     initializeDefaultCoreMemory — seeds
                                  the 19 fallback core sections, skip-
                                  if-existing so admin overrides win

Descriptions and skip/overwrite semantics preserved exactly — the two
paths use different description text ("from core.memory.json" vs
"(fallback)") so they stay as parallel code rather than collapsing
into a shared helper. All files under 215 lines.
